+++ resources/views/roles/index.blade.php
@@ -7,7 +7,7 @@
     <div class="panel panel-default">
         <div class="panel-heading clearfix">
             <p class="" style="">User Roles Manager</p>
-            @can('role-create')
+            @can('Role-create')
                 <a class="btn btn-success pull-right" style="margin-right:8px;" href="{{ route('roles.create') }}"><i class="glyphicon glyphicon-plus margin-right-sm"></i>&nbsp;New Role</a>
             @endcan
         </div>
@@ -25,14 +25,14 @@
                         <td>{{ $role->name }}</td>
                         <td>
                             <a class="btn btn-info" href="{{ route('roles.show',$role->id) }}">Show</a>
-{{--                            @can('role-edit')--}}
+                            @can('Role-edit')
                                 <a class="btn btn-primary" href="{{ route('roles.edit',$role->id) }}">Edit</a>
-{{--                            @endcan--}}
-{{--                            @can('role-delete')--}}
+                            @endcan
+                            @can('Role-delete')
                                 {!! Form::open(['method' => 'DELETE','route' => ['roles.destroy', $role->id],'style'=>'display:inline']) !!}
                                 {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
                                 {!! Form::close() !!}
-{{--                            @endcan--}}
+                            @endcan
                         </td>
                     </tr>
                 @endforeach
